(unistd.h): Move Win32 unistd.h out of src
Commits 4f0b369 and bf159cb were not tested on Linux, and caused
the build to fail on Linux. The reason is the supplied unistd.h is
only to be used for Win32 (MinGW and MSVC) builds, and conflicts
with existing headers on Linux. Relocate file to win32 include
directory, update Makefile.am to correctly include header in build.