(buildme): Add -l flag for script to disable quiet configure
Most times, we want configure to be quiet. However, there are times,
especially with CI builds, where bisecting a build fail requires
configure to be "loud." Add knob on the script to enables runtime
specification of non-quiet builds. This is particularly useful with
Travis, as a build can be triggered with a custom .travis.yml config,
and the '-l' flag could be turned on without needlessly poluting the
git commit history.