(configure.ac): Add clang as ObjC compiler, remove linking --tag
Commit 4ca18d4 was a work around for the observed error, which was
that when clang was used as the C compiler, linking would fail,
because gcc would be used as the linker. The explicit --tag=CC
within the src/Makefile.am overcame the glitch, but did not address
the root cause. Remove the unecessary --tag=CC in Makefile.am, and
add clang as a possible Objective C compiler.