(configure): Remove AC_PROG_CXX, use C compiler for Objective C
Project does not have any C++ code, remove the search and init from
configure.ac.

Project experiences problems when both gcc and clang are installed,
because the Objective-C compiler is used for final linking. This
was first problemmatic when the AC_PROG_OBJC only used gcc, but
clang was used for the rest of the code. The build would error out
with a note about being unable to infer tagged configuration. Adding
clang to the AC_PROG_OBJC cleared the error for clang builds, but
repeated the error on gcc builds. Now, the ObjC compiler will be
the C compiler sucessfully vetted by configure. This can be over-
ridden by user by specifying OBJC variable.