(buildme-latest): Create build script that fetches latest sources
This script is intended to be run as a cron job (eventually), to
check and catch errors in the dependent libraries (ffmpeg, giflib,
libpng, libjpeg and libexif). It uses CMake as the build mechanism
for several libraries. Edit the travis file to give the new script
a test run on wider machine set.